Moved UIUtils to a new folder
