+++ src/UI/UIUtils.java
@@ -1,4 +1,4 @@
-package src;
+package src.UI;
 
 import javax.swing.*;
 import java.awt.*;
